requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.41
pydantic==2.11.7
pydantic-settings==2.10.1
apscheduler==3.11.0
python-dotenv==1.1.1
httpx==0.28.1
google-genai
